chore: rename app everywhere
