+++ app/android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "work.mydata",
+    package = "com.egendata",
 )
 
 android_resource(
     name = "res",
-    package = "work.mydata",
+    package = "com.egendata",
     res = "src/main/res",
 )
