+++ app/android/app/src/main/java/com/egendata/MainActivity.java
@@ -1,4 +1,4 @@
-package work.mydata;
+package com.egendata;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "MyData";
+        return "Egendata";
     }
 }
