+++ app/android/app/src/main/java/com/egendata/MainApplication.java
@@ -1,4 +1,4 @@
-package work.mydata;
+package com.egendata;
 
 import android.app.Application;
 
@@ -31,6 +31,7 @@ public class MainApplication extends Application implements ReactApplication {
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
         new MainReactPackage(),
+            new AsyncStoragePackage(),
         new AsyncStoragePackage(),
         new RNSecureRandomPackage(),
         new SvgPackage(),
