+++ app/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "MyData",
-  "displayName": "MyData"
+  "name": "Egendata",
+  "displayName": "Egendata"
 }
