+++ app/ios/Egendata.xcodeproj/xcshareddata/xcschemes/Egendata-tvOS.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "MyData-tvOS.app"
-               BlueprintName = "MyData-tvOS"
-               ReferencedContainer = "container:MyData.xcodeproj">
+               BuildableName = "Egendata-tvOS.app"
+               BlueprintName = "Egendata-tvOS"
+               ReferencedContainer = "container:Egendata.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "MyData-tvOSTests.xctest"
-               BlueprintName = "MyData-tvOSTests"
-               ReferencedContainer = "container:MyData.xcodeproj">
+               BuildableName = "Egendata-tvOSTests.xctest"
+               BlueprintName = "Egendata-tvOSTests"
+               ReferencedContainer = "container:Egendata.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "MyData-tvOSTests.xctest"
-               BlueprintName = "MyData-tvOSTests"
-               ReferencedContainer = "container:MyData.xcodeproj">
+               BuildableName = "Egendata-tvOSTests.xctest"
+               BlueprintName = "Egendata-tvOSTests"
+               ReferencedContainer = "container:Egendata.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "MyData-tvOS.app"
-            BlueprintName = "MyData-tvOS"
-            ReferencedContainer = "container:MyData.xcodeproj">
+            BuildableName = "Egendata-tvOS.app"
+            BlueprintName = "Egendata-tvOS"
+            ReferencedContainer = "container:Egendata.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "MyData-tvOS.app"
-            BlueprintName = "MyData-tvOS"
-            ReferencedContainer = "container:MyData.xcodeproj">
+            BuildableName = "Egendata-tvOS.app"
+            BlueprintName = "Egendata-tvOS"
+            ReferencedContainer = "container:Egendata.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "MyData-tvOS.app"
-            BlueprintName = "MyData-tvOS"
-            ReferencedContainer = "container:MyData.xcodeproj">
+            BuildableName = "Egendata-tvOS.app"
+            BlueprintName = "Egendata-tvOS"
+            ReferencedContainer = "container:Egendata.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
