+++ app/ios/Egendata.xcodeproj/xcshareddata/xcschemes/Egendata.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "MyData.app"
-               BlueprintName = "MyData"
-               ReferencedContainer = "container:MyData.xcodeproj">
+               BuildableName = "Egendata.app"
+               BlueprintName = "Egendata"
+               ReferencedContainer = "container:Egendata.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "MyDataTests.xctest"
-               BlueprintName = "MyDataTests"
-               ReferencedContainer = "container:MyData.xcodeproj">
+               BuildableName = "EgendataTests.xctest"
+               BlueprintName = "EgendataTests"
+               ReferencedContainer = "container:Egendata.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "MyDataTests.xctest"
-               BlueprintName = "MyDataTests"
-               ReferencedContainer = "container:MyData.xcodeproj">
+               BuildableName = "EgendataTests.xctest"
+               BlueprintName = "EgendataTests"
+               ReferencedContainer = "container:Egendata.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "MyData.app"
-            BlueprintName = "MyData"
-            ReferencedContainer = "container:MyData.xcodeproj">
+            BuildableName = "Egendata.app"
+            BlueprintName = "Egendata"
+            ReferencedContainer = "container:Egendata.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "MyData.app"
-            BlueprintName = "MyData"
-            ReferencedContainer = "container:MyData.xcodeproj">
+            BuildableName = "Egendata.app"
+            BlueprintName = "Egendata"
+            ReferencedContainer = "container:Egendata.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "MyData.app"
-            BlueprintName = "MyData"
-            ReferencedContainer = "container:MyData.xcodeproj">
+            BuildableName = "Egendata.app"
+            BlueprintName = "Egendata"
+            ReferencedContainer = "container:Egendata.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
