+++ app/ios/Egendata/Egendata.entitlements
