+++ app/ios/Egendata/Images.xcassets/AppIcon.appiconset/Contents.json
