+++ app/ios/Egendata/Images.xcassets/AppIcon.appiconset/MyData2016-logo-1024x1024-20@2x.png
