+++ app/ios/Egendata/Images.xcassets/AppIcon.appiconset/MyData2016-logo-1024x1024-40@2x.png
