+++ app/ios/EgendataTests/EgendataTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface MyDataTests : XCTestCase
+@interface EgendataTests : XCTestCase
 
 @end
 
-@implementation MyDataTests
+@implementation EgendataTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
