+++ app/lib/screens/Home/Home.js
@@ -33,7 +33,7 @@ const HomeScreen = ({ navigation }) => {
   return (
     <Wrap style={{ justifyContent: 'flex-start', paddingHorizontal: 26 }}>
       <ScrollViewWrap contentContainerStyle={{ justifyContent: 'flex-start' }}>
-        <H1 style={{ marginTop: 64 }}>MyData</H1>
+        <H1 style={{ marginTop: 64 }}>Egendata</H1>
         <ConsentList consents={consents} />
       </ScrollViewWrap>
     </Wrap>
