+++ app/package.json
@@ -24,7 +24,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@egendata/messaging": "file:egendata-messaging-0.2.0.tgz",
-    "@react-native-community/async-storage": "1.4.0",
+    "@react-native-community/async-storage": "1.4.2",
     "@types/js-base64": "^2.3.1",
     "axios": "^0.18.0",
     "buffer": "^5.2.1",
